Use setuptools as the build backend, rather than flit
I think we can, and should, eventually switch to flit -- because flit
is faster and stricter, and because it catches more mistakes such as
when you build packages on a source tree with uncommitted changes.

But I have painted myself into a corner by the module naming
convention "src/sdx/datamodel" rather than "src/sdxdatamodel".  Flit
expects the the top-level module name to follow project name.
Setuptools is a little more forgiving, so using setuptools as build
backend for now will help us make progress.

Flit's strictness here is better because it also makes sure that
editable installs will continue to work.  With the former module
naming scheme, we have "src/sdx/datamodel" here, and "src/sdx/pce" in
pce, and because of that, we can't have editable installs in pce...